requirements.txt
streamlit==1.41.1
tensorflow==2.18.0
numpy==2.0.2
pillow==11.0.0